requirements.txt
streamlit
openai
Pillow
PyPDF2